+++ c/array-patterns/array22_pattern.c
@@ -29,7 +29,7 @@ void __VERIFIER_assert(int cond) { if(!(cond)) { ERROR: __VERIFIER_error(); } }
 extern int __VERIFIER_nondet_int() ;
 extern short __VERIFIER_nondet_short() ;
 
-short ARR_SIZE ;
+signed long long ARR_SIZE ;
 
 int outerBoundaryPatternCheck(short index)
 {
@@ -41,7 +41,7 @@ int outerBoundaryPatternCheck(short index)
 
 int main()
 {
-	ARR_SIZE = __VERIFIER_nondet_short() ;
+	ARR_SIZE = (signed long long)__VERIFIER_nondet_short() ;
 	__VERIFIER_assume(ARR_SIZE > 1) ;
 
 	int array[ARR_SIZE][ARR_SIZE] ;
